Split src/runs/stream.ts into stream/ subdirectory (warren-041e / pl-9088 step 5)
Extract event-bridge, terminal-detection, state-polling, cost-stats
persistence, and active-stream recovery into focused modules.

  src/runs/stream/
    bridge.ts            — bridgeRunStream pump
    terminal-detect.ts   — detectRuntimeTerminal, isPiAgentEnd
    run-state-poller.ts  — burrow runs.get fallback (warren-6596)
    stats.ts             — pi + claude cost persistence
    recover.ts           — recoverActiveRunStreams
    types.ts             — shared interfaces incl. SessionStats
    index.ts             — public re-exports

Tests split into bridge.test.ts, run-state-poller.test.ts, stats.test.ts,
recover.test.ts, plus terminal-detect.test.ts (moved from
src/runs/stream.detect.test.ts), backed by test-helpers.ts.

Callers updated to import from ./stream/index.ts (manage.ts, cancel.ts,
runs/index.ts, usage-aggregate.ts, reap/types.ts).

- file-size budgets for stream.{ts,test.ts} removed (all new files ≤ 500
  lines; largest is stats.test.ts at 497).
- biome noExcessiveLinesPerFunction override for stream.test.ts removed.
- noExcessiveCognitiveComplexity override migrated from stream.ts to
  stream/bridge.ts (carry-over — same single offending function).
