+++ .mulch/expertise/build.jsonl
@@ -86,3 +86,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T17:52:39.645Z","evidence":{"seeds":"warren-325d","commit":"89f2e888738ad5cc2432dc17b65259c31774c5d4"},"tags":["release","changelog","openapi"],"relates_to":["mx-92eaee"],"content":"Release patch flow (warren-325d / pl-5516 step 3) per .claude/commands/release.md: edit package.json version + src/index.ts VERSION in lockstep (release.yml fails if they disagree), move [Unreleased] entries to a new dated section in CHANGELOG.md, then run 'bun run gen:openapi' — docs/openapi.yaml's info.version is sourced from package.json, so gen:openapi:check fails check:all on drift. No CLAUDE.md/README.md edits needed for a pure patch with no command-surface change. Also: a fresh workspace needs 'bun install' before 'bun run check:all' — otherwise module-resolving tests fail (js-yaml, @os-eco/burrow-cli, @os-eco/plot-cli, drizzle-orm/bun-sqlite) and drag function coverage to ~83%, below the 85% floor.","id":"mx-bc5079"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T19:23:39.075Z","evidence":{"seeds":"warren-48de","commit":"302275b2510a128b03ba6912fe3ab60395333cdb"},"content":"Handlers monolith split phase 1 (warren-48de / pl-9088 step 1): src/server/handlers.ts → src/server/handlers/index.ts; plot handlers extracted to src/server/handlers/plots.ts (15 handlers covering /plots/*, /brainstorm), plot-plan-run handler to src/server/handlers/plot-plan-runs.ts. Pattern for the remaining 11 plan-9088 steps: (1) keep ROUTE_TABLE in handlers/index.ts as the canonical wire surface — gen:docs (scripts/generate-docs.ts HANDLERS_PATH) and gen:openapi (scripts/generate-openapi.ts) parse the array literal from one file. (2) Shared parsing helpers (readJsonBody, readJsonBodyOrEmpty, requireString, optionalString, requireParam, defaultSpawn) are exported from handlers/index.ts and re-imported by extracted modules — circular import is safe because handler bodies reference helpers lazily at request time, not at module evaluation. (3) Relative imports inside handlers/index.ts shift ../X → ../../X and ./X → ../X (siblings under src/server moved up one level). (4) ROUTE_TABLE imports the extracted handler functions from ./plots.ts and ./plot-plan-runs.ts. (5) Touch points: biome.json complexity-rule overrides, scripts/file-size-budgets.json, src/server/server.ts + src/server/index.ts + src/server/handlers.preview.test.ts + src/preview/proxy.test.ts all reference ./handlers/index.ts. (6) AGENTS.md grep-validates backticked paths so the handlers.ts → handlers/index.ts reference must be updated there too. (7) biome formatter may reformat single-line imports into multi-line when ../../ paths push them past 100 chars — budget for a small line-count uptick after biome --write and trim doc-comments to compensate so the file-size ratchet doesn't fail. (8) Plot-related test files moved alongside (handlers.plots.test.ts → handlers/plots.test.ts) with bulk sed-update of ../X → ../../X and ./X → ../X; inline dynamic imports (await import('../plots/index.ts')) need the same shift.","id":"mx-3df5c5"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-27T20:03:55.922Z","evidence":{"commit":"b66f19b23a3aa212e0054d2726758dd8ffe34a2c"},"relates_to":["mx-99ad0d"],"dir_anchors":["src/server/handlers"],"content":"Handlers monolith split phase 3 (warren-599c / pl-9088 step 3): src/server/handlers/index.ts → agents.ts (203 lines), projects.ts (261), burrows.ts (116), workers.ts (89), diagnostics.ts (136), meta.ts (42); handlers/index.ts shrinks from 1194 to 420 lines and now contains only the ROUTE_TABLE composer + shared parsing helpers + API_PREFIXES/isAuthExempt/buildApiRoutes. Pattern follows phase 1 (mx-3df5c5) and phase 2 (mx-99ad0d): shared helpers re-imported via './index.ts'; ROUTE_TABLE stays in handlers/index.ts as canonical wire surface for gen:docs and gen:openapi. Route handler symbol renames (listAgents→listAgentsHandler, healthz→healthzHandler, readyz→readyzHandler, version→versionHandler) propagate through docs/http-api.md + docs/openapi.yaml — must run 'bun run gen:docs && bun run gen:openapi' after the split. Tests reorganised: handlers.test.ts → handlers/agents.test.ts + handlers/projects.test.ts + handlers/brainstorm.test.ts; handlers.preview.test.ts → handlers/index.test.ts (isAuthExempt only) + handlers/meta.test.ts (/preview/config) + handlers/runs.preview.test.ts (login + teardown + proxy preamble); handlers.burrows.test.ts and handlers.workers.test.ts move under handlers/ without splitting. agents.ts handles BOTH GET/POST /agents routes AND POST /projects/:id/agents/refresh since projectCanopyClient + refreshProjectAgents are the agents subsystem. diagnostics.ts owns /readyz with private check helpers; meta.ts holds inert /healthz + /version + /preview/config probes. Removed handlers/index.ts from biome's noExcessiveCognitiveComplexity override list; added handlers/projects.ts (runProjectTriggerHandler stays complex at 22).","id":"mx-929bbc"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-27T20:40:35.382Z","evidence":{"seeds":"warren-041e","commit":"a6c6a58164f2e3d58eb431548d59b1b2a5fd6bb4"},"relates_to":["mx-929bbc"],"dir_anchors":["src/runs"],"content":"Stream bridge split (warren-041e / pl-9088 step 5): src/runs/stream.ts (911 lines) → src/runs/stream/ subdirectory. Modules: bridge.ts (bridgeRunStream pump, retains noExcessiveCognitiveComplexity override at score 63), terminal-detect.ts (detectRuntimeTerminal + isPiAgentEnd), run-state-poller.ts (warren-6596 fallback), stats.ts (pi+claude cost persistence: snapshotStats/persistPiStatsDelta/persistInStreamUsage), recover.ts (recoverActiveRunStreams + ActiveBridge types), types.ts (shared interfaces incl. SessionStats so usage-aggregate can re-export without runtime cycle). Tests split into bridge.test.ts (event flow + terminal detection + b1a9 ghost-run), run-state-poller.test.ts (warren-6596), stats.test.ts (PiStatsClient + in-stream pi + in-stream claude), recover.test.ts, plus terminal-detect.test.ts (moved from src/runs/stream.detect.test.ts), backed by test-helpers.ts (makePool, evt, source, piAgentEnd, piTurnEnd, claudeResult, seedBridgeRun). Callers updated to import from ./stream/index.ts (manage.ts, cancel.ts, runs/index.ts, usage-aggregate.ts, reap/types.ts). file-size budgets for stream.{ts,test.ts} removed; biome noExcessiveLinesPerFunction override for stream.test.ts removed. All files now ≤ 500 lines (largest: stats.test.ts at 497).","id":"mx-c7ca2b"}
