+++ scripts/file-size-budgets.json
@@ -34,8 +34,6 @@
 		"src/runs/pr.ts": 659,
 		"src/runs/spawn.test.ts": 1344,
 		"src/runs/spawn.ts": 882,
-		"src/runs/stream.test.ts": 1248,
-		"src/runs/stream.ts": 911,
 		"src/server/bridges.test.ts": 647,
 		"src/server/bridges.ts": 655,
 		"src/server/handlers/agents.test.ts": 533,
